+++ toxcore/group_moderation.h
@@ -38,7 +38,7 @@ struct GC_Ban {
     IP_Port     ip_port;
     uint8_t     nick[MAX_GC_NICK_SIZE];
     uint16_t    nick_len;
-    uint16_t    id;
+    uint32_t    id;
 };
 
 /* Holds data pertaining to a peer who has been banned or demoted to observer. */
@@ -62,7 +62,7 @@ struct GC_Sanction {
  * Returns length of unpacked data on success.
  * Returns -1 on failure.
  */
-int mod_list_unpack(GC_Chat *chat, const uint8_t *data, uint32_t length, uint16_t num_mods);
+int mod_list_unpack(GC_Chat *chat, const uint8_t *data, uint32_t length, uint32_t num_mods);
 
 /* Packs moderator list into data.
  * data must have room for num_mods * GC_MOD_LIST_ENTRY_SIZE bytes.
@@ -115,7 +115,7 @@ void mod_list_cleanup(GC_Chat *chat);
  * Returns -1 on failure.
  */
 int sanctions_list_pack(uint8_t *data, uint16_t length, struct GC_Sanction *sanctions,
-                        struct GC_Sanction_Creds *creds, uint16_t num_sanctions);
+                        struct GC_Sanction_Creds *creds, uint32_t num_sanctions);
 
 /* Unpack max_sanctions sanctions from data into sanctions, and unpacks credentials into creds.
  * Put the length of the data processed in processed_data_len.
@@ -123,7 +123,7 @@ int sanctions_list_pack(uint8_t *data, uint16_t length, struct GC_Sanction *sanc
  * Returns number of unpacked entries on success.
  * Returns -1 on failure.
  */
-int sanctions_list_unpack(struct GC_Sanction *sanctions, struct GC_Sanction_Creds *creds, uint16_t max_sanctions,
+int sanctions_list_unpack(struct GC_Sanction *sanctions, struct GC_Sanction_Creds *creds, uint32_t max_sanctions,
                           const uint8_t *data, uint16_t length, uint16_t *processed_data_len);
 
 /* Packs sanction list credentials into data.
@@ -154,7 +154,7 @@ int sanctions_list_make_creds(GC_Chat *chat);
  * Returns -1 if one or more entries are invalid.
  */
 int sanctions_list_check_integrity(const GC_Chat *chat, struct GC_Sanction_Creds *creds,
-                                   struct GC_Sanction *sanctions, uint16_t num_sanctions);
+                                   struct GC_Sanction *sanctions, uint32_t num_sanctions);
 
 /* Adds an entry to the sanctions list. The entry is first validated and the resulting
  * new sanction list is compared against the new credentials.
@@ -192,14 +192,14 @@ int sanctions_list_remove_observer(GC_Chat *chat, const uint8_t *public_key, str
  * Returns 0 on success.
  * Returns -1 on failure or if entry was not found
  */
-int sanctions_list_remove_ban(GC_Chat *chat, uint16_t ban_id, struct GC_Sanction_Creds *creds);
+int sanctions_list_remove_ban(GC_Chat *chat, uint32_t ban_id, struct GC_Sanction_Creds *creds);
 
 /* Replaces all sanctions list signatures made by public_sig_key with the caller's.
  * This is called whenever the founder demotes a moderator.
  *
  * Returns the number of entries re-signed.
  */
-uint16_t sanctions_list_replace_sig(GC_Chat *chat, const uint8_t *public_sig_key);
+uint32_t sanctions_list_replace_sig(GC_Chat *chat, const uint8_t *public_sig_key);
 
 /* Creates a new sanction list hash and puts it in hash.
  *
@@ -208,8 +208,8 @@ uint16_t sanctions_list_replace_sig(GC_Chat *chat, const uint8_t *public_sig_key
  *
  * If num_sanctions is 0 the hash is zeroed.
  */
-void sanctions_list_make_hash(struct GC_Sanction *sanctions, uint32_t new_version,
-                              uint16_t num_sanctions, uint8_t *hash);
+void sanctions_list_make_hash(struct GC_Sanction *sanctions, uint32_t new_version, uint32_t num_sanctions,
+                              uint8_t *hash);
 
 void sanctions_list_cleanup(GC_Chat *chat);
 
@@ -222,26 +222,26 @@ void sanctions_list_cleanup(GC_Chat *chat);
 bool sanctions_list_ip_banned(const GC_Chat *chat, IP_Port *ip_port);
 
 /* Returns the number of sanctions list entries that are of type SA_BAN */
-uint16_t sanctions_list_num_banned(const GC_Chat *chat);
+uint32_t sanctions_list_num_banned(const GC_Chat *chat);
 
 /* Fills list with all valid ban ID's. */
-void sanctions_list_get_ban_list(const GC_Chat *chat, uint16_t *list);
+void sanctions_list_get_ban_list(const GC_Chat *chat, uint32_t *list);
 
 /* Returns the nick length of the ban entry associted with ban_id on success.
  * Returns 0 if ban_id does not exist.
  */
-uint16_t sanctions_list_get_ban_nick_length(const GC_Chat *chat, uint16_t ban_id);
+uint16_t sanctions_list_get_ban_nick_length(const GC_Chat *chat, uint32_t ban_id);
 
 /* Copies the nick associated with ban_id to nick.
  *
  * Returns 0 on success.
  * Returns -1 if ban_id does not exist.
  */
-int sanctions_list_get_ban_nick(const GC_Chat *chat, uint16_t ban_id, uint8_t *nick);
+int sanctions_list_get_ban_nick(const GC_Chat *chat, uint32_t ban_id, uint8_t *nick);
 
 /* Returns a timestamp indicating when the ban designated by ban_id was set.
  * Returns 0 if ban_id does not exist.
  */
-uint64_t sanctions_list_get_ban_time_set(const GC_Chat *chat, uint16_t ban_id);
+uint64_t sanctions_list_get_ban_time_set(const GC_Chat *chat, uint32_t ban_id);
 
 #endif /* GROUP_MODERATION_H */
